add resnet_dtan_center_loss

+++ prcv_expreiment/trainging/dtgn_resnet_train_center_loss.py
@@ -30,7 +30,7 @@ def placeholder_inputs():
     #     images_placeholders.append(tf.placeholder(tf.float32, shape=[None, 64, 64, 1]))
     images_placeholder = tf.placeholder(tf.float32, shape=[None, 64, 64, model.OULU_SIMPLE_NUM])
     # landmarks_placeholder = tf.placeholder(tf.float32, shape=[None, LANDMARK_LENGTH])
-    labels_placeholder = tf.placeholder(tf.int32, shape=(None, model.OULU_NUM_CLASSES))
+    labels_placeholder = tf.placeholder(tf.float32, shape=(None, model.OULU_NUM_CLASSES))
     keep_prob = tf.placeholder("float")
     is_train = tf.placeholder(tf.bool, name='phase_train')
     return images_placeholder, labels_placeholder, keep_prob, is_train
